flashtripdemo/competitor/requirements.txt
aiocontextvars==0.2.1
aiohttp==3.4.4
aioredis==1.1.0
async-timeout==3.0.1
asyncpg==0.18.2
attrs==19.1.0
chardet==3.0.4
contextvars==2.3
gino==0.8.1
hiredis==1.0.0
idna==2.8
idna-ssl==1.1.0
immutables==0.6
lxml==4.3.3
motor==2.0.0
multidict==4.5.2
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2018.4
six==1.12.0
SQLAlchemy==1.3.3
tornado==6.0.2
yarl==1.2.3
